Make interpreting output value of models easier
Change preprocessing so now output is related to the most central value
insted of being centered around it. It will make models output
percentage difference instead of absolute difference what is more
robust and versatile in case of different economic factors that
could influence the accuracy of the model.

Add _out_norm_params variable to all of the models. Now each model
remembers how the output it was trained on was normalized so it is
possible to revert the normalization process and get the interpretable
output. It was not possible before.